+++ src/main/java/me/vineer/economyapi/money/Balance.java
@@ -22,29 +22,29 @@ public class Balance {
     public static final NamespacedKey MoneyType = new NamespacedKey(EconomyAPI.getPlugin(), "type");
 
 
-    private int money;
-    private int donateMoney;
+    private double money;
+    private double donateMoney;
     private final String member;
 
-    private Balance(int money, int donateMoney, String member) {
+    private Balance(double money, double donateMoney, String member) {
         this.money = money;
         this.donateMoney = donateMoney;
         this.member = member;
     }
-    public void setMoney(int money) {
+    public void setMoney(double money) {
         this.money = money;
         setPlayerBalance(this.money, donateMoney, member);
     }
 
-    public int getMoney() {
+    public double getMoney() {
         return money;
     }
 
-    public int getDonateMoney() {
+    public double getDonateMoney() {
         return donateMoney;
     }
 
-    public void setDonateMoney(int donateMoney) {
+    public void setDonateMoney(double donateMoney) {
         this.donateMoney = donateMoney;
         setPlayerBalance(money, this.donateMoney, member);
     }
@@ -54,7 +54,7 @@ public class Balance {
     }
 
 
-    public boolean sendMoneyTo(String toSendPlayer, int amount) {
+    public boolean sendMoneyTo(String toSendPlayer, double amount) {
         Balance toBalance = Balance.getPlayerBalance(toSendPlayer);
         if(getMoney() - amount < 0)return false;
         setMoney(getMoney() - amount);
@@ -62,12 +62,12 @@ public class Balance {
         return true;
     }
 
-    public static void setPlayerBalance(int money, int donateMoney, String member) {
+    public static void setPlayerBalance(double money, double donateMoney, String member) {
         if(hasBalance(member)) {
             try {
                 PreparedStatement pr = Database.getConnection().prepareStatement("UPDATE Players SET money = ?, donateMoney = ? WHERE Name = ?");
-                pr.setInt(1, money);
-                pr.setInt(2, donateMoney);
+                pr.setDouble(1, money);
+                pr.setDouble(2, donateMoney);
                 pr.setString(3, member);
                 pr.executeUpdate();
             } catch (SQLException e) {
@@ -77,8 +77,8 @@ public class Balance {
             try {
                 PreparedStatement pr = Database.getConnection().prepareStatement("INSERT INTO Players(Name, money, donateMoney) VALUES (?, ?, ?)");
                 pr.setString(1, member);
-                pr.setInt(2, money);
-                pr.setInt(3, donateMoney);
+                pr.setDouble(2, money);
+                pr.setDouble(3, donateMoney);
                 pr.executeUpdate();
             } catch (SQLException e) {
                 e.printStackTrace();
@@ -94,7 +94,7 @@ public class Balance {
             ResultSet rs = pr.executeQuery();
             has_account = rs.next();
             if(has_account) {
-                return new Balance(rs.getInt("money"), rs.getInt("donateMoney"), player);
+                return new Balance(rs.getDouble("money"), rs.getDouble("donateMoney"), player);
             }
         } catch (SQLException e) {
             e.printStackTrace();
@@ -106,12 +106,12 @@ public class Balance {
         return new Balance(300, 0, player);
     }
 
-    public static void changePlayerBalance(String member, int money, int donateMoney) {
+    public static void changePlayerBalance(String member, double money, double donateMoney) {
         if(hasBalance(member)) {
             try {
                 PreparedStatement pr = Database.getConnection().prepareStatement("UPDATE Players SET money = money + ?, donateMoney = donateMoney + ? WHERE Name = ?");
-                pr.setInt(1, money);
-                pr.setInt(2, donateMoney);
+                pr.setDouble(1, money);
+                pr.setDouble(2, donateMoney);
                 pr.setString(3, member);
                 pr.executeUpdate();
             } catch (SQLException e) {
@@ -134,7 +134,7 @@ public class Balance {
         }
         return has_account;
     }
-    public static ItemStack createCheck(MoneyType type, int amount, String fromPlayerDisplay, String fromPlayer) {
+    public static ItemStack createCheck(MoneyType type, double amount, String fromPlayerDisplay, String fromPlayer) {
         ItemStack item = null;
         if(type.equals(me.vineer.economyapi.money.MoneyType.MONEY)) {
             item = ItemCreator.createGuiItem(Material.PAPER, ChatColor.WHITE + "Чек", ChatColor.GREEN + "$" + ChatColor.WHITE + amount, ChatColor.WHITE + "от " + fromPlayerDisplay);
@@ -144,18 +144,18 @@ public class Balance {
             Balance.changePlayerBalance(fromPlayer, 0, -amount);
         }
         ItemMeta itemMeta = item.getItemMeta();
-        itemMeta.getPersistentDataContainer().set(Moneykey, PersistentDataType.INTEGER, amount);
+        itemMeta.getPersistentDataContainer().set(Moneykey, PersistentDataType.DOUBLE, amount);
         itemMeta.getPersistentDataContainer().set(MoneyType, PersistentDataType.STRING, type.getName());
         item.setItemMeta(itemMeta);
         return item;
     }
-    public static void changeCheck(@NotNull ItemStack check, MoneyType type, Integer amount, String fromPlayer) {
+    public static void changeCheck(@NotNull ItemStack check, MoneyType type, Double amount, String fromPlayer) {
         ItemMeta meta = check.getItemMeta();
         if(type != null) {
             meta.getPersistentDataContainer().set(MoneyType, PersistentDataType.STRING, type.getName());
         }
         if(amount != null) {
-            meta.getPersistentDataContainer().set(Moneykey, PersistentDataType.INTEGER, amount);
+            meta.getPersistentDataContainer().set(Moneykey, PersistentDataType.DOUBLE, amount);
             if(meta.getPersistentDataContainer().get(MoneyType, PersistentDataType.STRING).equals(me.vineer.economyapi.money.MoneyType.MONEY.getName())) {
                 List<String> lore = meta.getLore();
                 lore.set(0, ChatColor.GREEN + "$" + ChatColor.WHITE + amount);
